Can now loop over multiple sessions if they are in metadata.
This distinguishes a new session, defined in
session_params.py:read_metadata() from old sessions.

A new session reads the metadata and the whole pipeline will have
to be run (minus LFP maybe).

If something changes, maybe only a part of the pipeline needs to
be run, for instance the aligned traces and data object creation.
In that case, the new block will load the metadata for all
sessions in the lists and loop over the pipeline.